+++ mvc/src/main/java/nextstep/mvc/controller/exception/NotSupportHandler.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.controller.tobe.exception;
+package nextstep.mvc.controller.exception;
 
 public class NotSupportHandler extends RuntimeException {
